rsx: Move SPIRV generation into common code
